requirements.txt
ultralytics== 8.3.5
streamlit~=1.39.0
opencv-python-headless
Pillow~=10.4.0
numpy~=2.1.2
matplotlib

opencv-python~=4.10.0.84